create common keccak file with sponge tools
